% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacusPlot.R
\name{abacusPlot}
\alias{abacusPlot}
\title{Plot daily detections over time (or across recievers)}
\usage{
abacusPlot(ATTdata, id = NULL, theme = "theme_linedraw", xlab = NULL,
  ylab = NULL, det.col = 2, tag.col = 8, facet = FALSE,
  new.window = TRUE, ...)
}
\arguments{
\item{ATTdata}{input 'ATT' object containing tag detection, metadata and station information.}

\item{id}{input Tag.ID for a single or several tags within the 'ATT' object to plot}

\item{theme}{ggplot theme used in plot (default = "theme_linedraw")}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{det.col}{color of points indicating detections (default = "red")}

\item{tag.col}{color of points indicating start and end of tag life (default = "grey")}

\item{facet}{conditional argument to plot faceted plot, plotting detections by each animal on receiver stations (default = FALSE)}

\item{new.window}{conditional argument for plotting in a new window (default = TRUE)}

\item{...}{other arguments sent to geom_points to modify plotting aesthetics}
}
\value{
Produces an abacus plot in a new window. Daily detections over time for each tag
  (determined by "Tag.ID" field name). If facet parameter is TRUE detections of each tag on
  receiver stations plotted.
}
\description{
Produces an abacus plot to visualise patterns of detection over time or across signal receivers
}
\examples{
## Import example datasets
data(IMOSdata)
data(taginfo)
data(statinfo)

## Setup data
ATTdata<- setupData(Tag.Detections = IMOSdata, 
                    Tag.Metadata = taginfo, 
                    Station.Information = statinfo, 
                    source="IMOS")

## Create abacus plot
abacusPlot(ATTdata)


}
\seealso{
Input data needs to be setup using \code{\link{setupData}}
}
