% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COA.R
\name{COA}
\alias{COA}
\title{Short-term Centers of Activity}
\usage{
COA(ATTdata, timestep = 60, split = FALSE)
}
\arguments{
\item{ATTdata}{an 'ATT' object created using \code{\link{setupData}} containing
tag detection data, metadata and station information}

\item{timestep}{an integer containing the temporal bin size of center of activity calculations (in minutes), default 60 min}

\item{split}{a conditional argument to split COA estimates by Tag.ID. Default is FALSE, which produces
a single tibble with COAs from all individuals compiled. If TRUE, a list of tibbles produced with each Tag.ID seperated.}
}
\value{
Exports a 'COA' object that is a tibble (if split = FALSE) or a list of tibbles (if split = TRUE)
}
\description{
Function to calculate Short-term centers of activity positions from passive telemetry data.
 Based on technique described in: Simpfendorfer, C. A., M. R. Heupel, and R. E. Hueter. 2002.
 Estimation of short-term centers of activity from an array of omnidirectional hydrophones and its
 use in studying animal movements. Canadian Journal of Fisheries and Aquatic Sciences 59:23-32.
}
\examples{
## Import example datasets
data(IMOSdata)
data(taginfo)
data(statinfo)

## Setup data
ATTdata <- setupData(Tag.Detections = IMOSdata, 
                     Tag.Metadata = taginfo, 
                     Station.Information = statinfo, 
                     source="IMOS")

## Estimate Short-term Center of Activities for all individuals
COAdata<- COA(ATTdata)
head(COAdata)

}
\seealso{
Input data needs to be setup using \code{\link{setupData}}. 'COA' object required for \code{\link{HRSummary}}
}
                                                                                 