\name{ComputeAzimuth}
\alias{ComputeAzimuth}
\title{
Compute the Azimuth Between Two Coordinates
}
\description{
This function computes the Azimuth from two geographical coordinates. These locations must be in decimal degrees. 
}
\usage{
ComputeAzimuth(lat1, lat2, lon1, lon2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lat1 }{
the latitude of the first coordinate 
}
  \item{lat2 }{
the latitude of the second coordinate
}
  \item{lon1 }{
the longitude of the first coordinate
}
  \item{lon2 }{
the longitude of the second coordinate
}
}
\details{
Coordinates are given in decimal degrees (WGS 84)
}
\keyword{spatial}