% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionSummary.R
\name{detectionSummary}
\alias{detectionSummary}
\title{Calculate standard metrics of detection using passive telemetry data}
\usage{
detectionSummary(ATTdata, sub = "\%Y-\%m")
}
\arguments{
\item{ATTdata}{an 'ATT' object created using \code{\link{setupData}} containing
tag detection data, metadata and station information}

\item{sub}{argument to define temporal subsets. Temporal subsets are currently restricted to monthly ('\%Y-\%m') or weekly ('\%Y-\%W').
Defaults to monthly if none is provided.}
}

\item{download_date}{date the array was downloaded in the form 'yyyy-mm-dd' for calculating detection summary before the collection of full dataset has been collected.
If none provided, it is assumed complete data has been collected, and detection indices are calculated as such.}
}
\value{
Produces a list of tibbles containing Overall (full tag life) and Subsetted (user-defined temporal subsets) metrics of detection.
  Temporal subsets are currently restricted to monthly ('\%Y-\%m') or weekly ('\%Y-\%W').
}
\description{
Produce standard metrics of detection. Metrics include Number of detections, Number of receiver stations
 detected on, Number of days detected, Number of Days at Liberty and Detection Index.
}
\examples{
## Import example datasets
data(IMOSdata)
data(taginfo)
data(statinfo)

## Setup data
ATTdata <- setupData(Tag.Detections = IMOSdata,
                     Tag.Metadata = taginfo, 
                     Station.Information = statinfo, 
                     source="IMOS")

## Estimate detecion metrics with monthly subsets chosen
#dSum <- detectionSummary(ATTdata, sub = "%Y-%m")

## Metrics of detection for full tag life
#dSum$Overall

## Metrics of detection for each temporal subset
#dSum$Subsetted

}
\seealso{
Input data needs to be setup using \code{\link{setupData}}
}
