% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersalSummary.R
\name{dispersalSummary}
\alias{dispersalSummary}
\title{Calculate metrics of dispersal using passive telemetry data}
\usage{
dispersalSummary(ATTdata)
}
\arguments{
\item{ATTdata}{an 'ATT' object created using \code{\link{setupData}} containing
tag detection data, metadata and station information}
}
\value{
Produces a tibble containing dispersal distance and bearings between tag release location and
  each detection as well as between consecutive detections.
}
\description{
Produce standard metrics of dispersal. Metrics include dispersal distance and bearings between
 tag release location and each detection as well as between consecutive detections.
}
\examples{
## Import example datasets
data(IMOSdata)
data(taginfo)
data(statinfo)

## Setup data
ATTdata<- setupData(Tag.Detections = IMOSdata, 
                    Tag.Metadata = taginfo, 
                    Station.Information = statinfo, 
                    source = "IMOS")

## Estimate detection metrics with monthly subsets chosen
dispSum <- dispersalSummary(ATTdata)

}
\seealso{
Input data needs to be setup using \code{\link{setupData}}
}
