% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSummary.R
\name{HRSummary}
\alias{HRSummary}
\title{Calculate metrics of activity space using passive telemetry data}
\usage{
HRSummary(COAdata, projCRS, type = "MCP", cont = c(50, 95),
  sub = "\%Y-\%m", cumulative = FALSE, storepoly = FALSE, h = 500,
  ext = 2, grid = 200, div = 4)
}
\arguments{
\item{COAdata}{a 'COA' object with estimated center of activity positions (using \code{\link{COA}})}

\item{projCRS}{a coordinate reference system in meters used to accurately estimate area (see \code{\link{CRS}})}

\item{type}{type of activity space metric to calculate. Currently "MCP" for minimum convex polygons,
"fKUD" for fixed KUD and "BBKUD" for Brownian bridge KUD. Defaults to "MCP" when no value provided.}

\item{cont}{contours of activity space models to estimate areas. Defaults to 50\% and 95\% contours}

\item{sub}{temporal subset used to calculate subsetted activity space metrics.
Currently supports monthly ("\%Y-\%m") or weekly ("\%Y-\%W"). Defaults to monthly.}

\item{cumulative}{TRUE/FALSE, should the operation calculate cumulative activity space areas. Caution
this may take a long time depending on size of dataset.}

\item{storepoly}{TRUE/FALSE, should activity space metrics be saved as spatial objects (polygons or rasters)}

\item{h}{smoothing factor (in m) associated with error associated with reciever range of passive telemetry system.
Defaults to 200 m if none provided.}

\item{ext}{spatial extent used to calculate probability density metrics of activity space ("fKUD", "BBKUD")}

\item{grid}{grid size used to calculate probability density metrics of activity space ("fKUD", "BBKUD")}

\item{div}{sig1 divisor used to correct BBKUD estimates.}
}
\value{
Produces a list of 2 tibbles containing Overall (full tag life) and Subsetted (user-defined temporal subsets) metrics of activity space.
  If storepoly=TRUE additional object within list containing spatial objects (MCP polygons or KUD rasters).
  Temporal subsets are currently restricted to monthly ("\%Y-\%m") or weekly ("\%Y-\%W"). Cumulative measures across temporal subsets
  included if cumulative=TRUE.
}
\description{
Produce standard metrics of activity space. Metrics can include Minimum convex polygon area (MCP),
 fixed Kernel Utilisation Distribution (fKUD) area and Brownian Bridge KUDs (BBKUD). Activity space metrics are
 calculated for full period of tag life and user defined subset (currently monthly or weekly only). Cumulative
 activity spaces can also be estimated at set temporal subsets.
}
\examples{

require(sp)

## Import example datasets
data(IMOSdata)
data(taginfo)
data(statinfo)

## Setup data
ATTdata <- setupData(Tag.Detections = IMOSdata,
                     Tag.Metadata = taginfo, 
                     Station.Information = statinfo, 
                     source = "IMOS")

## Estimate short-term center of activities
COAdata <- COA(ATTdata)

## Define coordinate system for projection of detection data (in m)
proj <- CRS("+init=epsg:3577")

## Estimate 100\% MCP areas
mcp_est <- HRSummary(COAdata, projCRS=proj, type="MCP", cont=100)

## Estimate 50\% and 95\% fKUD areas with cumulative metrics calculated
#kud_est <- HRSummary(COAdata, 
#                    projCRS=proj,
#                    type="fKUD", 
#                    cumulative=TRUE)

## Estimate 20\%, 50\% and 95\% BBKUD contour areas and store polygons
#kud_est <- HRSummary(COAdata, 
#                     projCRS=proj, 
#                     type="BBKUD", 
#                     cont=c(20,50,95), 
#                     storepoly=TRUE)


}
\seealso{
Input data needs to be setup using \code{\link{setupData}}, and COAs calculated using \code{\link{COA}}.
}
