% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupData.R
\name{setupData}
\alias{setupData}
\title{Setup data prior to use with Animal Tracking Toolbox functions}
\usage{
setupData(Tag.Detections, Tag.Metadata, Station.Information, source = NULL, tzone = "UTC",
  crs = NULL)
}
\arguments{
\item{Tag.Detections}{data frame with tag detection data including coordinates. Currently only handles IMOS ATF data structure.}

\item{Tag.Metadata}{data frame with metadata for all tags represented in Tag.Detections. Currently only handles IMOS ATF metadata structure.}

\item{Station.Information}{data frame with information on receiver station including coordinates. Currently only handles IMOS ATF station information structure.}

\item{source}{source character indicating source of Tag.Detection data. "IMOS" for data downloaded from IMOS data repository and "VEMCO" for data exported from the VEMCO VUE database}

\item{tzone}{time zone of date time information in Tag.Detections, Tag.Metadata and Station.Information. If none provided defaults to "UTC"}

\item{crs}{geographic coordinate system for all Tag.Detections, Tag.Metadata and Station.Information (latitude/longitude). If none provided defaults to WGS84.}
}
\value{
Produces an 'ATT' object that is a list of tibbles containing Tag.Detections, Tag.Metadata and Station.Information.
  The 'ATT' object will have a geographic coordinate system associated with it for smoother functioning of subsequent functions.
}
\description{
Sets up and stores all relevant information required for spatio-temporal analyses of passive telemetry data from IMOS and VEMCO databases.
This function produces an 'ATT' object which standardises field names (currently handles IMOS ATF export data structure).
}
\examples{
## Import example datasets
data(IMOSdata)
data(VEMCOdata)
data(taginfo)
data(statinfo)

## Setup data from IMOS ATF data repository
ATTdata <- setupData(Tag.Detections = IMOSdata,
                     Tag.Metadata = taginfo, 
                     Station.Information = statinfo, 
                     source = "IMOS")
head(ATTdata)

## Setup data from VEMCO VUE database
ATTdata <- setupData(Tag.Detections = VEMCOdata, 
                     Tag.Metadata = taginfo, 
                     Station.Information = statinfo, 
                     source = "VEMCO")
head(ATTdata)

}
\seealso{
setup data can be used to estimate detection \code{\link{detectionSummary}}, dispersal \code{\link{dispersalSummary}}
  and Short-term center of activity \code{\link{COA}}.
}
