\name{VTrack-package}
\alias{VTrack-package}
\alias{VTrack}
\docType{package}
\title{
VTrack: A package housing a collection of R Functions for the Analysis of Passive Acoustic Telemetry Data Collected by VEMCO Acoustic Tracking Equipment
}
\description{
The \code{VTrack} package was built to facilitate the assimilation, analysis and synthesis of animal location data collected by the VEMCO suite of acoustic transmitters and receivers. As well as database and visualisation capabilities, VTrack also provides functions to identify ecologically relevant events from tag detection and sensor data (e.g. residence, movements, diving, surfacing events). This procedure condenses the acoustic detection database by orders of magnitude, greatly enhancing the synthesis of acoustic detection data.

https://cran.r-project.org/web/packages/VTrack/index.html

VTrack version >2.0.0 now hosts the Animal Tracking Toolbox (ATT). The ATT calculates standardised metrics of dispersal and activity space to enable direct comparisons between animals tracked within the same study and between studies or locations. The functions use tag detection data alongside tag metadata and receiver station information to calculate standardised movement and activity space metrics. 

}
\details{
\tabular{ll}{
Package: \tab VTrack\cr
Type: \tab Package\cr
Version: \tab 2.0.0\cr
Date: \tab 2018-12-05\cr
License: \tab GPL(>=2)\cr
}

}
\author{
Ross Dwyer, Mathew Watts, Craig Franklin 
School of Biological Sciences, University of Queensland, StLucia, Queensland, Australia.

Hamish Campbell
Charles Darwin UNiversity

Vinay Udyawer
Australian Institute for Marine Sciences

Maintainer: Ross Dwyer <ross.dwyer@uq.edu.au>

}
\references{

To cite the 'VTrack' package in publications:

Campbell HA, Watts ME, Dwyer RG, Franklin CE (2012) V-Track: software for analysing and visualising animal movement from acoustic telemetry detections, Marine and Freshwater Research 63:815-820

To cite the ATT toolbox within the VTrack package:
Udyawer V, Dwyer RG, Hoenner X, Babcock RC, Brodie S, Campbell HA, Harcourt RG, Huveneers C, Jaine FRA, Simpfendorfer CA, Taylor MD, Heupel MR (in press) A standardised framework for analysing animal detections from automated tracking arrays. Animal Biotelemetry doi
}

\keyword{ package }
