\name{VTrack-package}
\alias{VTrack-package}
\alias{VTrack}
\docType{package}
\title{
VTrack: A Collection of Tools for the Analysis of Remote Acoustic Telemetry Data
}
\description{
The package \code{VTrack} was designed by researchers at the University of Queensland to allow the analysis and visualisation of data generated from the VEMCO suite of passive and active acoustic receivers.
}
\details{
\tabular{ll}{
Package: \tab VTrack\cr
Type: \tab Package\cr
Version: \tab 1.4.1\cr
Date: \tab 2018-02-19\cr
License: \tab GPL(>=2)\cr
}

}
\author{
Ross Dwyer, Mathew Watts, Hamish Campbell & Craig Franklin E.C.O.-lab and the E.D.G., School of Biological Sciences, University of Queensland, StLucia, Queensland, Australia.

Maintainer: Ross Dwyer <ross.dwyer@uq.edu.au>

}
\references{

Campbell, H.A., Watts, M.E., Dwyer, R.G., Franklin, C.E. 2012. V-Track: software for analysing and visualising animal movement from acoustic telemetry detections. Marine and Freshwater Research, 63:815-820.}

\keyword{ package }
